+++ src/common/CalcTemplates.c
