+++ src/common/ardopSampleArrays.c
