+++ src/common/ardopcommon.h
